Widget_state and layout renames
