+++ ipywidgets/index.js
@@ -13,8 +13,8 @@ require('underscore');
 var register = require("./static/widgets/js/register");
 [
     require("./static/widgets/js/manager-base"),
-    require("./static/widgets/js/layout"),
     require("./static/widgets/js/widget"),
+    require("./static/widgets/js/widget_layout"),
     require("./static/widgets/js/widget_link"),
     require("./static/widgets/js/widget_bool"),
     require("./static/widgets/js/widget_button"),
