+++ ipywidgets/static/index.js
@@ -4,12 +4,12 @@
 define([
     "base/js/namespace",
     "./widgets/js/register",
-    "./widgets/js/widget_state",
+    "./widgets/js/save_state",
     "./widgets/js/manager-base",
     "./widgets/js/manager",
     "./widgets/js/utils",
     "./widgets/js/widget",
-    "./widgets/js/layout",
+    "./widgets/js/widget_layout",
     "./widgets/js/widget_link",
     "./widgets/js/widget_bool",
     "./widgets/js/widget_button",
